Implement some of the filesystem
This commit marks the start of actually providing a filesystem for with
fuse to mount. Initial mounting as well as listing the root directory
works. You should also be able to `stat` files and the root directory.
mtime, atime and ctime are currently broken and default to UNIX_EPOCH.

What the filesystem does not do is actually fetching files and folders
as requested. It just errors out.
Next up is figuring out when to fetch file-metadata from WebDAV. And
actually downloading files.